refactor: Merged preprocessor into core
